Split animation page from map construction page
